refactor: snake position depends on game path.
